LayoutTests/fast/multicol/balance-trailing-border2.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
    description("Tests that the bottom border at a column boundary in a block that lives in multiple columns is recorded properly for balancing");
</script>
<p>There should be no space between the short black line and the long gray one.</p>
<div id="mc" style="-webkit-columns:3; columns:3; line-height:21px; orphans:1; widows:1; border-bottom:3px solid gray;">
    <div style="height:9px;"></div>
    <div style="border-bottom:3px solid black;">
        <br><br><br><br><br><br><br><br>
    </div>
</div>
<script>
    shouldBe("mc.offsetHeight", "69");
</script>
